requirements.txt
requests
python-dotenv
sqlalchemy
flask
pytest-mock
pytest